requirements.txt
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
async-timeout==4.0.3
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
decorator==5.1.1
dill==0.3.7
distlib==0.3.7
Django==4.2.4
django-redis==5.3.0
djangorestframework==3.14.0
executing==1.2.0
filelock==3.12.2
identify==2.5.26
idna==3.4
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
nodeenv==1.8.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
pre-commit==3.3.3
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.12.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.2
wcwidth==0.2.6
wrapt==1.15.0
